Use activeWindowChanges, activeViewComponentChanges
The Codecov extension still used workspace.onDidOpenTextDocument, which can be subject to race conditions (the event can be fired before the Codecov extension is activated). This is changed to use `workspace.activeWindowChanges` and `Window.activeViewComponentChanges` when available, to reliably decorate editors.

Also makes sure that all subscriptions are correctly handled through context.Subscriptions.